+++ Tests/Integrationtest.Config.ps1
@@ -3,7 +3,7 @@ $ModuleRoot = Split-Path (Resolve-Path "$ProjectRoot\*\*.psm1")
 $ModuleName = Split-Path $ModuleRoot -Leaf
 Import-Module (Join-Path "$ModuleRoot" "$ModuleName.psm1")
 
-$Global:connectionString = 'url=http://localhost/AppServerAR82Test/;AcceptSelfSigned=true;AllowServerNameMismatch=true'
+$Global:connectionString = 'url=http://XXX/AppServer/;AcceptSelfSigned=true;AllowServerNameMismatch=true'
 $Global:factory = 'QBM.AppServer.Client.ServiceClientFactory'
-$Global:authenticationString = 'Module=DialogUser;User=viadmin;Password='
-$Global:modulesToSkip = 'AAD','ACN','ADS','APC','ATT','CAP','CCC','CPL','CSM','DPR','EBS','EX0','EXH','GAP','HDS','LDP','NDO','O3E','O3S','PAG','POL','RMB','RMS','RPS','SAC','SAP','SBW','SHR','SP0','TSB','UCI','UNX'
+$Global:authenticationString = 'Module=DialogUser;User=cccadmin;Password='
+$Global:modulesToSkip = $null #'AAD','ACN','ADS','APC','ATT','CAP','CCC','CPL','CSM','DPR','EBS','EX0','EXH','GAP','HDS','LDP','NDO','O3E','O3S','PAG','POL','RMB','RMS','RPS','SAC','SAP','SBW','SHR','SP0','TSB','UCI','UNX'